pyproject.toml
[tool.poetry]
name = "core-engine"
version = "0.0.1"
description = ""
authors = ["EasyGDS"]

[tool.poetry.dependencies]
python = "3.8.10"
marshmallow = "3.2.1"
flask-cors = "3.0.8"
pyjwt = "1.7.1"
cryptography = "^35.0.0"
python-slugify = "^5.0.2"
requests = "2.24.0"
click = "^8.0.3"
sentry-sdk = {extras = ["flask"], version = "^1.4.3"}
psycopg2-binary = "2.8.6"
SQLAlchemy = "1.3.19"
alembic = "1.7.5"
ipython = "^7.29.0"
celery = "^5.2.3"
PyYAML = "^6.0"
pymongo = "^4.0.1"
redis = "^4.1.2"
gevent = "^21.12.0"
semantic-version = "^2.9.0"
uWSGI = "^2.0.20"
eventlet = "^0.33.0"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
